+++ scripts/linux_package.py
@@ -1,7 +1,5 @@
-import json
-import os
+import platform
 import shlex
-import shutil
 import subprocess
 import xml.etree.ElementTree as ET
 
@@ -16,6 +14,7 @@ def call(cmd: str, capture_output = True):
 
 data = ET.parse("Amethyst/Amethyst.csproj").getroot()
 version = str(data.find("PropertyGroup").find("Version").text)  # type: ignore
+arch = platform.uname().machine
 
-for i in ["deb", "pacman", "apk"]:
-    call(f'fpm -s dir -t {i} --name amethyst --license MIT --version {version} --description "A programming language for Minecraft data packs" --url "https://www.amethyst.dev" --maintainer "kinderhead" -a native Amethyst/dist/amethyst=/usr/bin/amethyst Amethyst/dist/core=/usr/share/amethyst', False)
+for (pkg, ext) in [("deb", ".deb"), ("pacman", ".pkg.tar.zst"), ("apk", ".apk")]:
+    call(f'fpm -s dir -t {pkg} --name amethyst --license MIT --version {version} --description "A programming language for Minecraft data packs" --url "https://www.amethyst.dev" --maintainer "kinderhead" -a {arch} -p amethyst-{version}-{arch}{ext} Amethyst/dist/amethyst=/usr/bin/amethyst Amethyst/dist/std=/usr/share/amethyst', False)