+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * Fixed an issue with forking `execute` statements not actually forking.
 * Fixed an issue where scopes and redefinitions weren't processed correctly.
 * Changed the MacOS bundle to use tarballs.
+* Moved `core` to `std/core`.
+* Allow standard library searching in `/usr/share/amethyst/std`.